chore(oui-navbar): update navigation for new components
