+++ packages/oui-navbar/src/group/navbar-group.controller.js
@@ -1,11 +1,57 @@
 export default class {
-    constructor ($attrs, $element, ouiNavbarConfiguration, NavbarGroupService) {
+    constructor ($attrs, $element, ouiNavbarConfiguration, KEYBOARD_KEYS) {
         "ngInject";
 
         this.$attrs = $attrs;
         this.$element = $element;
         this.config = ouiNavbarConfiguration;
-        this.navbarGroupService = NavbarGroupService;
+
+        this.KEYBOARD_KEYS = KEYBOARD_KEYS;
+    }
+
+    bindGroup (groupName) {
+        const keys = {};
+        const keysRegex = new RegExp([
+            this.KEYBOARD_KEYS.ALT,
+            this.KEYBOARD_KEYS.TAB
+        ].join("|"));
+
+        const tabbableItems = this.navbarCtrl.getGroup(groupName);
+        const lastIndex = tabbableItems.length - 1;
+        const focusElement = (e, groupIndex) => {
+            let index = groupIndex;
+            keys[e.which] = true;
+
+            if (keys[this.KEYBOARD_KEYS.ALT] && !keys[this.KEYBOARD_KEYS.TAB]) {
+                // Move Down
+                index = index >= lastIndex ? 0 : index + 1;
+            } else if (keys[this.KEYBOARD_KEYS.ALT] && keys[this.KEYBOARD_KEYS.TAB]) {
+                // Move Up
+                index = index <= 0 ? lastIndex : index - 1;
+            }
+
+            // Check if element is visible
+            if (tabbableItems[index].clientHeight) {
+                tabbableItems[index].focus();
+            } else {
+                focusElement(e, index);
+            }
+        };
+
+        angular.element(tabbableItems)
+            .on("keydown", (e) => {
+                if (keysRegex.test(e.which) && this.isOpen(groupName)) {
+                    e.preventDefault();
+                    focusElement(e, this.navbarCtrl.getGroup(groupName).indexOf(e.target));
+                }
+            })
+            .on("keyup", (e) => {
+                delete keys[e.which];
+            });
+    }
+
+    isOpen (state) {
+        return this.navbarCtrl.navigation && this.navbarCtrl.navigation[state];
     }
 
     $onInit () {
@@ -16,11 +62,11 @@ export default class {
     }
 
     $postLink () {
-        this.navbarGroupService.addItemToGroup(this.$element[0], this.groupName);
+        this.navbarCtrl.addItemToGroup(this.$element[0], this.groupName);
 
         // Bind items when it's the last item
         if (this.isLast) {
-            this.navbarGroupService.bindGroup(this.groupName);
+            this.bindGroup(this.groupName);
         }
     }
 }
