+++ packages/oui-navbar/src/group/navbar-group.directive.js
@@ -2,6 +2,9 @@ import controller from "./navbar-group.controller";
 
 export default () => ({
     restrict: "A",
+    require: {
+        navbarCtrl: "^ouiNavbar"
+    },
     bindToController: {
         groupName: "@ouiNavbarGroup",
         isLast: "<?ouiNavbarGroupLast"
