+++ packages/oui-navbar/src/menu/navbar-menu.html
@@ -15,12 +15,11 @@
     </span>
     <h3 class="oui-navbar-menu__title" ng-bind="::$ctrl.headerTitle"></h3>
 </header>
-<ul class="oui-navbar-list" ng-if="::$ctrl.expanded">
+<ul class="oui-navbar-list">
     <li class="oui-navbar-list__item {{::!!menuLink.class ? menuLink.class : null}}"
         ng-repeat="menuLink in ::$ctrl.menuLinks track by $index"
-        ng-class="::{
-            'oui-navbar-menu__item_active': !!menuLink.isActive
-        }">
+        ng-class="::{ 'oui-navbar-menu__item_active': !!menuLink.isActive }"
+        ui-sref-active="oui-navbar-menu__item_active">
 
         <!-- Link (href) -->
         <a class="oui-navbar-link"
@@ -32,9 +31,8 @@
                 'oui-navbar_desktop-only': !!menuLink.subLinks,
                 'oui-navbar-link_external': !!menuLink.isExternal
             }"
-            ng-click="$ctrl.toggleMenu()"
+            ng-click="$ctrl.navbarCtrl.toggleMenu()"
             ng-href="{{::menuLink.url}}"
-            ng-init="::$last && $ctrl.setFirstFocus()"
             oui-navbar-group="{{::$ctrl.menuName}}"
             oui-navbar-group-last="::$last">
             <span ng-bind="::menuLink.title"></span>
@@ -44,7 +42,7 @@
         </a>
         <!-- /Link (href) -->
 
-        <!-- /Link (ui-sref) -->
+        <!-- Link (ui-sref) -->
         <a class="oui-navbar-link"
             ng-if="::!!menuLink.state"
             ng-attr-aria-label="{{::!!menuLink.label ? menuLink.label : null}}"
@@ -52,10 +50,9 @@
             ng-class="::{
                 'oui-navbar_desktop-only': !!menuLink.subLinks
             }"
-            ng-click="$ctrl.toggleMenu()"
+            ng-click="$ctrl.navbarCtrl.toggleMenu()"
             ng-bind="::menuLink.title"
             ui-sref="{{::$ctrl.constructor.getFullSref(menuLink)}}"
-            ng-init="::$last && $ctrl.setFocus()"
             oui-navbar-group="{{::$ctrl.menuName}}"
             oui-navbar-group-last="::$last">
         </a>
@@ -68,7 +65,6 @@
             ng-attr-title="{{::!!menuLink.label ? menuLink.label : null}}"
             ng-click="$ctrl.closeMenuWithCallback($event, menuLink.click)"
             ng-bind="::menuLink.title"
-            ng-init="::$last && $ctrl.setFocus()"
             oui-navbar-group="{{::$ctrl.menuName}}"
             oui-navbar-group-last="::$last">
         </button>
@@ -76,16 +72,15 @@
 
         <!-- Button for Next Level -->
         <button class="oui-navbar-link" type="button"
-            aria-haspopup="true" aria-expanded="{{!!$ctrl.navigation[menuLink.name]}}"
+            aria-haspopup="true" aria-expanded="{{!!$ctrl.navbarCtrl.navigation[menuLink.name]}}"
             ng-if="::!!menuLink.subLinks && !!menuLink.subLinks.length"
             ng-attr-aria-label="{{::!!menuLink.label ? menuLink.label : null}}"
             ng-attr-title="{{::!!menuLink.label ? menuLink.label : null}}"
             ng-class="::{
                 'oui-navbar_mobile-only': !!menuLink.url || !!menuLink.state
             }"
-            ng-click="$ctrl.toggleMenu(menuLink.name, true)"
+            ng-click="$ctrl.navbarCtrl.toggleMenu(menuLink.name, true)"
             ng-bind="::menuLink.title"
-            ng-init="::$last && $ctrl.setFocus()"
             oui-navbar-group="{{::$ctrl.menuName}}"
             oui-navbar-group-last="::$last">
         </button>
@@ -94,15 +89,13 @@
         <!-- Menu for Next Level -->
         <oui-navbar-menu class="oui-navbar_mobile-only"
             ng-if="::!!menuLink.subLinks && !!menuLink.subLinks.length"
-            ng-class="::!!$ctrl.childrenClass ? $ctrl.childrenClass : null"
-            children-class="{{::$ctrl.childrenClass}}"
             back-button="true"
             header-breadcrumb="{{::menuLink.headerBreadcrumb || $ctrl.getChildBreadcrumb()}}"
             header-title="{{::menuLink.headerTitle || menuLink.title}}"
             name="{{::menuLink.name}}"
             links="::menuLink.subLinks"
-            navigation="$ctrl.navigation"
-            expanded="$ctrl.navigation[menuLink.name]">
+            align="{{::$ctrl.align}}"
+            fixed="::$ctrl.fixed">
         </oui-navbar-menu>
         <!-- Menu for Next Level -->
