+++ packages/oui-navbar/src/navbar.controller.js
@@ -1,7 +1,7 @@
 import { addBooleanParameter } from "@oui-angular/common/component-utils";
 
 export default class {
-    constructor ($attrs, $document, $element, $timeout, ouiNavbarConfiguration, NavbarService, KEYBOARD_KEYS) {
+    constructor ($attrs, $document, $element, $timeout, ouiNavbarConfiguration, KEYBOARD_KEYS) {
         "ngInject";
 
         this.$attrs = $attrs;
@@ -9,13 +9,56 @@ export default class {
         this.$element = $element;
         this.$timeout = $timeout;
         this.config = ouiNavbarConfiguration;
-        this.navbarService = NavbarService;
+
+        this.keyboardNav = {};
         this.KEYBOARD_KEYS = KEYBOARD_KEYS;
     }
 
+    addItemToGroup (item, groupName) {
+        // Create the group if it doesn't exist
+        if (angular.isUndefined(this.keyboardNav[groupName])) {
+            this.keyboardNav[groupName] = [];
+        }
+
+        // Add item to the group
+        this.keyboardNav[groupName].push(item);
+    }
+
+    getGroup (groupName) {
+        return this.keyboardNav[groupName];
+    }
+
+    // Set focus to an item of a group
+    setFocusTo (groupName, index) {
+        // Add a delay to force focus
+        const delay = 50;
+        this.$timeout(() => this.keyboardNav[groupName][index] && this.keyboardNav[groupName][index].focus(), delay);
+    }
+
     toggleMenu (state, isInternalNav) {
-        // Update navbar navigation
-        this.navigation = this.navbarService.toggleMenu(state, isInternalNav);
+        if (state) {
+            // Reset navigation if not internal navigation
+            if (!isInternalNav && (!this.navigation || !this.navigation[state])) {
+                this.navigation = {};
+            }
+
+            if (isInternalNav || !this.navigation[state]) {
+                // Toggle menu if internal navigation or state is undefined
+                this.navigation[state] = !this.navigation[state];
+            } else if (this.navigation[state]) {
+                // Else close all menus
+                this.navigation = null;
+            }
+
+            // Focus first list item when opened
+            if (this.navigation && this.navigation[state]) {
+                // Add a little delay to avoid transition bug on Webkit
+                this.setFocusTo(state, 0);
+            }
+        } else {
+            // Close all menus
+            this.navigation = null;
+        }
     }
 
     updateLinks () {
