+++ packages/oui-navbar/src/navbar.html
@@ -1,7 +1,7 @@
 <!-- Toggler -->
 <div ng-transclude="togglerSlot">
     <oui-navbar-toggler
-        ng-if="::$ctrl.togglerLinks"
+        ng-if="::!!($ctrl.$attrs['togglerLinks'] || !!$ctrl.$attrs['mainLinks'])"
         links="::$ctrl.togglerLinks">
     </oui-navbar-toggler>
 </div>
@@ -10,7 +10,7 @@
 <!-- Brand -->
 <div ng-transclude="brandSlot">
     <oui-navbar-brand
-        ng-if="::$ctrl.brand"
+        ng-if="::!!$ctrl.$attrs['brand']"
         href="{{::$ctrl.brand.url}}"
         heading="{{::$ctrl.brand.title}}"
         aria-label="{{::$ctrl.brand.label}}"
@@ -23,9 +23,9 @@
 
 <!-- Main Links -->
 <div class="oui-navbar-list oui-navbar-list_main oui-navbar_desktop-only"
-    ng-if="::$ctrl.mainLinks"
     ng-transclude="mainSlot">
     <oui-navbar-link
+        ng-if="::!!$ctrl.$attrs['mainLinks']"
         name="{{::managerLink.name}}"
         text="{{::managerLink.title}}"
         href="{{::managerLink.url}}"
@@ -39,18 +39,18 @@
 
 <!-- Aside Links -->
 <div class="oui-navbar-list oui-navbar-list_aside oui-navbar-list_end"
-    ng-if="::$ctrl.asideLinks"
     ng-transclude="asideSlot">
 
     <!-- Internal -->
     <oui-navbar-dropdown
+        ng-if="::!!$ctrl.$attrs['asideLinks']"
         name="{{::asideLink.name}}"
         text="{{::asideLink.title}}"
         aria-label="{{::!!asideLink.label ? asideLink.label : null}}"
         icon-badge="(asideLink.subLinks | filter: {isActive: 'true', acknowledged: '!true'}).length"
         icon-class="{{::!!asideLink.iconClass ? asideLink.iconClass : null}}"
         on-click="asideLink.onClick"
-        ng-repeat="asideLink in ::$ctrl.asideLinks track by $index"
+        ng-repeat="asideLink in $ctrl.asideLinks track by $index"
         ng-class="::!!asideLink.class ? asideLink.class : null"
         ng-switch="::asideLink.name">
 
@@ -58,7 +58,7 @@
         <oui-navbar-notification ng-switch-when="notifications"
             class="oui-navbar-menu_fixed oui-navbar-menu_right"
             name="{{::asideLink.name}}"
-            links="::asideLink.subLinks"
+            links="asideLink.subLinks"
             limit-to="::asideLink.limitTo"
             header-template="::asideLink.headerTemplate"
             header-title="{{::asideLink.headerTitle || asideLink.title}}"
@@ -70,28 +70,24 @@
 
         <!-- User Menu -->
         <oui-navbar-menu ng-switch-when="user"
-            class="oui-navbar-menu_fixed oui-navbar-menu_user oui-navbar-menu_right"
-            children-class="oui-navbar-menu_fixed oui-navbar-menu_right"
             header-breadcrumb="{{::asideLink.nichandle}}"
             header-title="{{::asideLink.fullName}}"
             name="{{::asideLink.name}}"
             links="::asideLink.subLinks"
-            navigation="$ctrl.navigation"
-            expanded="$ctrl.navigation[asideLink.name]">
+            align="end"
+            fixed>
         </oui-navbar-menu>
         <!-- /User Menu -->
 
         <!-- Default Menu -->
         <oui-navbar-menu ng-switch-default
-            class="oui-navbar-menu_fixed oui-navbar-menu_right"
-            children-class="oui-navbar-menu_fixed oui-navbar-menu_right"
             header-class="oui-navbar_mobile-only"
             header-breadcrumb="{{::!!asideLink.headerBreadcrumb ? asideLink.headerBreadcrumb : null}}"
             header-title="{{::asideLink.headerTitle || asideLink.title}}"
             name="{{::asideLink.name}}"
             links="::asideLink.subLinks"
-            navigation="$ctrl.navigation"
-            expanded="$ctrl.navigation[asideLink.name]">
+            align="end"
+            fixed>
         </oui-navbar-menu>
         <!-- Default Menu -->
 
@@ -102,7 +98,7 @@
 <!-- Aside Links -->
 
 <!-- Backdrop -->
-<div class="oui-navbar-backdrop" aria-hidden="true"
+<div class="oui-navbar-backdrop" aria-hidden="true" tabindex="-1"
     ng-class="{ 'oui-navbar-backdrop_active': !!$ctrl.navigation }"
     ng-click="$ctrl.toggleMenu()">
 </div>
