+++ packages/oui-navbar/src/toggler/navbar-toggler.controller.js
@@ -1,0 +1,8 @@
+export default class {
+    $onChanges (changes) {
+        // Get links changes for the loader
+        if (changes.links) {
+            this.loaded = !!changes.links.currentValue;
+        }
+    }
+}
